TypeScript source!
This might be a slightly inadvisable development but I am experimenting.
The output is still JavaScript, but the input is now a TypeScript file.
Only a few very minor modifications had to be made so that it would
compile, and the output (which is now the target of Jasmine tests) still
functions as before, so there's that.